% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_project.R
\name{add_regions}
\alias{add_regions}
\title{Add Region Information to Project Data}
\usage{
add_regions(Project, regions)
}
\arguments{
\item{Project}{Project data frame}

\item{regions}{Regions data frame}
}
\value{
Project data frame with added region information
}
\description{
Helper function to add region information to Project data
}
\keyword{internal}
