% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_client.R
\name{Client_redact}
\alias{Client_redact}
\title{Redact PII from Client Data}
\usage{
Client_redact(Client, clients_to_filter = NULL)
}
\arguments{
\item{Client}{Client data frame}

\item{clients_to_filter}{Optional list of client IDs to filter out}
}
\value{
Redacted Client data
}
\description{
Redacts personally identifiable information while applying data quality rules
}
