% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_hud_export_utils.R
\name{extract_hud_export}
\alias{extract_hud_export}
\title{Extract HUD Export Archive}
\usage{
extract_hud_export(
  archive_path,
  extract_path = fs::path("data", "hmis"),
  delete_archive = TRUE,
  moment = Sys.Date(),
  wait = lubridate::minutes(2)
)
}
\arguments{
\item{archive_path}{Character. Path to the zip archive}

\item{extract_path}{Character. Directory where files should be extracted}

\item{delete_archive}{Logical. Whether to delete the archive after extraction}

\item{moment}{POSIXct/Date. Minimum acceptable file timestamp}

\item{wait}{Duration. Maximum time to wait for file}
}
\value{
Logical indicating success
}
\description{
Extract HUD Export Archive
}
\keyword{internal}
