% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_looks.R
\name{find_latest_looker_files}
\alias{find_latest_looker_files}
\title{Find the Most Recent Looker Files in S3}
\usage{
find_latest_looker_files(bucket, prefix = NULL, region = "us-east-2")
}
\arguments{
\item{bucket}{Character. Name of the S3 bucket}

\item{prefix}{Character. Prefix to filter S3 objects}

\item{region}{Character. AWS region}
}
\value{
List of information about the most recent files
}
\description{
Find the Most Recent Looker Files in S3
}
\keyword{internal}
