% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Geocodes}
\alias{Geocodes}
\title{Ohio Homeless Geocodes}
\format{
A data frame with 88 rows and 3 variables:
\describe{
\item{GeographicCode}{Code for the geography}
\item{State}{State abbreviation}
\item{Name}{City or county name}
\item{Type}{Type for name (City or County)}
\item{County}{County for the geography}
}
}
\source{
Ohio Balance of State Continuum of Care
}
\usage{
Geocodes
}
\description{
A dataset containing the mapping of Ohio counties to their respective
Geocodes
}
\keyword{datasets}
