% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmis_csv_specs.R
\docType{data}
\name{hmis_csv_specs}
\alias{hmis_csv_specs}
\title{HMIS Data Column Specifications}
\format{
An object of class \code{list} of length 24.
}
\usage{
hmis_csv_specs
}
\description{
Defines column specifications for different HMIS CSV files
}
\keyword{datasets}
