% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_disabilities.R
\name{load_disabilities}
\alias{load_disabilities}
\title{Load Disabilities Data}
\usage{
load_disabilities()
}
\arguments{
\item{extract_path}{Path where the HUD export files are located}
}
\value{
Disabilities data
}
\description{
Loads Disabilities data from CSV
}
