% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_enrollment.R
\name{load_enrollment}
\alias{load_enrollment}
\title{Load Enrollment as Enrollment_extra_Client_Exit_HH_CL_AaE}
\usage{
load_enrollment(
  Enrollment,
  EnrollmentCoC,
  Enrollment_extras,
  Exit,
  Client,
  Project,
  Referrals,
  rm_dates,
  app_env = get_app_env(e = rlang::caller_env())
)
}
\arguments{
\item{Enrollment_extras}{From Clarity Looker API Extras}
}
\description{
Load Enrollment as Enrollment_extra_Client_Exit_HH_CL_AaE
}
