% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_hud_csv.R
\name{load_hmis_csv}
\alias{load_hmis_csv}
\title{Load Specific HMIS CSV File from S3}
\usage{
load_hmis_csv(
  file_name,
  bucket = "hud.csv-daily",
  folder = "hmis_files",
  region = "us-east-2",
  col_types = NULL
)
}
\arguments{
\item{file_name}{\code{(character)} Name of the CSV file to load (without path)}

\item{bucket}{\code{(character)} Name of the S3 bucket}

\item{folder}{\code{(character)} Folder within the bucket where CSV files are stored}

\item{region}{\code{(character)} AWS region}
}
\value{
\code{(data.frame)} Data frame containing the loaded CSV data
}
\description{
Load Specific HMIS CSV File from S3
}
