% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_hmis_parquet.R
\name{load_hmis_parquet}
\alias{load_hmis_parquet}
\title{Load Specific HMIS Parquet File from S3}
\usage{
load_hmis_parquet(
  file_name,
  bucket = "hud.csv-daily",
  folder = "hmis_output",
  region = "us-east-2"
)
}
\arguments{
\item{file_name}{\code{(character)} Name of the parquet file to load (without path)}

\item{bucket}{\code{(character)} Name of the S3 bucket}

\item{folder}{\code{(character)} Folder within the bucket where parquet files are stored}

\item{region}{\code{(character)} AWS region}
}
\value{
\code{(data.frame)} Data frame containing the loaded parquet data
}
\description{
Load Specific HMIS Parquet File from S3
}
