% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_looks.R
\name{load_looker_data}
\alias{load_looker_data}
\title{Download Latest Looker CSV Files from S3}
\usage{
load_looker_data(
  bucket = "looker-daily",
  prefix = NULL,
  region = "us-east-2",
  delete_s3_object = FALSE,
  filename = NULL,
  col_types = NULL
)
}
\arguments{
\item{bucket}{Character. Name of the S3 bucket (default: "looker-daily")}

\item{prefix}{Character. Prefix to filter S3 objects}

\item{region}{Character. AWS region (default: "us-east-2")}

\item{delete_s3_object}{Logical. Whether to delete the object from S3 after download}

\item{local_path}{Character. Path where CSV files will be saved}
}
\value{
List of data frames containing the loaded CSV data
}
\description{
Download Latest Looker CSV Files from S3
}
