% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_project.R
\name{load_project_coc}
\alias{load_project_coc}
\title{Load Project CoC Data}
\usage{
load_project_coc()
}
\arguments{
\item{extract_path}{Path where the HUD export files are located}
}
\value{
Project CoC data
}
\description{
Loads Project CoC data from CSV
}
