% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_referrals.R
\name{load_referrals}
\alias{load_referrals}
\title{Load Referrals with Flexible Data Source}
\usage{
load_referrals(
  data = NULL,
  use_s3 = TRUE,
  bucket = "looker-daily",
  prefix = "CE_Referrals_extras",
  region = "us-east-2",
  local_path = fs::path("data", "looker"),
  delete_s3_object = FALSE,
  app_env = NULL,
  log_level = "INFO"
)
}
\arguments{
\item{data}{data.frame. Optional pre-loaded referrals data}

\item{app_env}{environment. Optional environment for storing results}

\item{log_level}{character. Level of logging detail}
}
\value{
Updated app_env or list of processed data
}
\description{
Loads and processes referral data with support for multiple input methods
}
