% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_services.R
\name{load_services}
\alias{load_services}
\title{Load Services Data}
\usage{
load_services()
}
\arguments{
\item{extract_path}{Path where the HUD export files are located}
}
\value{
Services data
}
\description{
Loads Services data from CSV
}
