% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/look_specs.R
\docType{data}
\name{look_specs}
\alias{look_specs}
\title{Looker Data Column Specifications}
\format{
An object of class \code{list} of length 13.
}
\usage{
look_specs
}
\description{
Defines column specifications for different custom Looks
}
\keyword{datasets}
