% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ServiceAreas}
\alias{ServiceAreas}
\title{Ohio Homeless Service Areas}
\format{
A data frame with 88 rows and 2 variables:
\describe{
\item{ssvf_service_area}{Name of the SSVF service}
\item{county}{County Name}
}
}
\source{
Ohio Balance of State Continuum of Care
}
\usage{
ServiceAreas
}
\description{
A dataset containing the mapping of Ohio counties to their respective
SSVF service
}
\keyword{datasets}
