% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_hmis_data.R
\name{upload_hmis_data}
\alias{upload_hmis_data}
\title{Upload Data Frame to S3 in Various Formats}
\usage{
upload_hmis_data(
  data,
  file_name,
  format = c("parquet", "feather", "csv", "rds"),
  bucket = "hud.csv-daily",
  folder = "hmis_output",
  region = "us-east-2",
  overwrite = TRUE
)
}
\arguments{
\item{data}{\code{(data.frame)} Data frame to upload}

\item{file_name}{\code{(character)} Base name for the uploaded file (without extension)}

\item{format}{\code{(character)} Format to save data in: "csv", "parquet", or "feather"}

\item{bucket}{\code{(character)} Name of the S3 bucket for upload}

\item{folder}{\code{(character)} Folder within the bucket where file will be stored}

\item{region}{\code{(character)} AWS region}

\item{overwrite}{\code{(logical)} Whether to overwrite if file exists}
}
\value{
\code{(logical)} TRUE if successful, FALSE otherwise
}
\description{
Upload Data Frame to S3 in Various Formats
}
