% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_hud_export.R
\name{load_hud_export}
\alias{load_hud_export}
\title{Download and Extract HUD Export from S3}
\usage{
load_hud_export(
  bucket = "hud.csv-daily",
  prefix = "HMIS",
  region = "us-east-2",
  s3_key = NULL,
  extract_path = fs::path("data", "hmis"),
  delete_archive = TRUE,
  delete_s3_object = FALSE,
  moment = Sys.Date()
)
}
\arguments{
\item{bucket}{\code{(character)} Name of the S3 bucket}

\item{s3_key}{\code{(character)} S3 object key (path) for the HUD export zip (optional)}

\item{extract_path}{\code{(character)} Path where the archive will be extracted}

\item{delete_archive}{\code{(logical)} Whether to delete the downloaded archive after extraction}

\item{delete_s3_object}{\code{(logical)} Whether to delete the object from S3 after download}

\item{moment}{\code{(POSIXct/Date)} Time point to validate archive recency}
}
\value{
\code{(logical)} Success of download and extraction
}
\description{
Download and Extract HUD Export from S3
}
