% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_project.R
\name{load_project}
\alias{load_project}
\title{Load and Process Project Data}
\usage{
load_project(extract_path = fs::path("data", "hmis"), regions = NULL)
}
\arguments{
\item{extract_path}{Path where the HUD export files are located}

\item{regions}{Optional data frame containing region definitions}
}
\value{
Processed Project data with CoC information
}
\description{
Loads Project data from CSV and processes it with related CoC data
}
